Address PR comments.

+++ freud/box.pyx
@@ -364,7 +364,7 @@ cdef class Box:
             l_masses = freud.util._convert_array(masses, shape=(len(vecs), ))
             l_masses_ptr = &l_masses[0]
 
-        cdef unsigned int Np = l_points.shape[0]
+        cdef size_t Np = l_points.shape[0]
         cdef vec3[float] result = self.thisptr.centerOfMass(
             <vec3[float]*> &l_points[0, 0], Np, l_masses_ptr)
         return np.asarray([result.x, result.y, result.z])
@@ -392,7 +392,7 @@ cdef class Box:
             l_masses = freud.util._convert_array(masses, shape=(len(vecs), ))
             l_masses_ptr = &l_masses[0]
 
-        cdef unsigned int Np = l_points.shape[0]
+        cdef size_t Np = l_points.shape[0]
         self.thisptr.center(<vec3[float]*> &l_points[0, 0], Np, l_masses_ptr)
         return vecs
